adding a sign up hook with navigation to login
