+++ src/services/users.ts
@@ -17,15 +17,18 @@ const signup = async (props: ISignup) => {
     .then(async response => {
       if (response.status === 201) {
         showMessage('success', 'success')
+        return true;
       } else if (response.status === 400) {
         showMessage('error', 'email already exists\nplease try another one');
+        return false;
       } else {
         showMessage('error', 'an unexpected error occured')
-        // throw new Error('Unexpected response status');
+        throw new Error('Unexpected response status');
       }
     })
     .catch(error => {
       showMessage('error', error);
+      return false
     });
 };
 export { signup };
